pyproject.toml
[tool.poetry]
name = "packer-cli"
version = "0.1.0"
description = "CLI for Sourced Packer Framework"
authors = ["Russell Whelan <russell.whelan@devopsgroup.com>"]
readme = "README.md"
packages = [
    {include = "packer_cli"},
    {include = "packer_cli/templates"}
]

[tool.poetry.dependencies]
python = "^3.11"
boto3 = "^1.28.53"
click = "^8.1.7"
jinja2 = "^3.1.2"
pyyaml = "^6.0.1"
click-logging = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
packer-cli = "packer_cli.cmd:cli"
